Move and redeclare ShareBroadcastReceiver
