+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
             android:value="@integer/google_play_services_version" />
 
         <receiver
-            android:name=".utils.ShareBroadcastReceiver"
+            android:name=".receivers.ShareBroadcastReceiver"
             android:label="ShareBroadcastReceiver" />
 
         <receiver
